requirements.txt
# Telegram Bot
python-telegram-bot>=22.2

# PostgreSQL
asyncpg>=0.31.0

# Environment
python-dotenv>=1.2.1

# Logging
loguru>=0.7.3

google-api-python-client==2.150.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

cyrtranslit==1.2.0